All OO classes in a same file
Implemented Luca's correction: all classes for different energies can be defined in the same file. I kept 2 different files for FCCee and CEPC for convenience.
